Cosmetic changes

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: pleander <pleander@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 14:35:12 by pleander          #+#    #+#             */
-/*   Updated: 2024/04/25 13:50:30 by pleander         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:29:43 by pleander         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <stddef.h>
 
 size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
